fix: process data in chunks for bounding box to not run out of memory
